--- imp_data.py
@@ -1,53 +1,0 @@
-redPath=[(0,4), (1,4), (1,3), (1,2), (1,1), (2,1), (3,1), (4,1), (5,1), (6,1), (7,1), (7,2), (7,3), (7,4), (7,5), (7,6), (7,7), (6,7), (5,7), (4,7), (3,7), (2,7), (1,7), (1,6), (1,5), 
-            (2,6), (3,6), (4,6), (5,6), (6,6), (6,5), (6,4), (6,3), (6,2), (5,2), (4,2), (3,2), (2,2), (2,3), (2,4), (2,5), 
-                (3,5), (4,5), (5,5), (5,4), (5,3), (4,3), (3,3),(3,4), 
-                    (4,4)]
-
-
-bluePath=[(4,0), (4,1), (5,1), (6,1), (7,1), (7,2), (7,3), (7,4), (7,5), (7,6), (7,7), (6,7), (5,7), (4,7), (3,7), (2,7), (1,7), (1,6), (1,5), (1,4), (1,3), (1,2), (1,1), (2,1), (3,1), 
-             (2,2), (2,3), (2,4), (2,5), (2,6), (3,6), (4,6), (5,6), (6,6), (6,5), (6,4), (6,3), (6,2), (5,2), (4,2), (3,2), 
-                 (3,3),(3,4), (3,5), (4,5), (5,5), (5,4), (5,3), (4,3),
-                    (4,4)]
-
-
-yellowPath=[(8,4), (7,4), (7,5), (7,6), (7,7), (6,7), (5,7), (4,7), (3,7), (2,7), (1,7), (1,6), (1,5), (1,4), (1,3), (1,2), (1,1), (2,1), (3,1), (4,1), (5,1), (6,1), (7,1), (7,2), (7,3), 
-             (6,2), (5,2), (4,2), (3,2), (2,2), (2,3), (2,4), (2,5), (2,6), (3,6), (4,6), (5,6), (6,6), (6,5), (6,4), (6,3), 
-                 (5,3), (4,3),(3,3),(3,4), (3,5), (4,5), (5,5), (5,4), 
-                    (4,4)]
-
-
-greenPath=[(4,8), (4,7), (3,7), (2,7), (1,7), (1,6), (1,5), (1,4), (1,3), (1,2), (1,1), (2,1), (3,1), (4,1), (5,1), (6,1), (7,1), (7,2), (7,3), (7,4), (7,5), (7,6), (7,7), (6,7), (5,7),
-              (6,6), (6,5), (6,4), (6,3), (6,2), (5,2), (4,2), (3,2), (2,2), (2,3), (2,4), (2,5), (2,6), (3,6), (4,6), (5,6), 
-                (5,5), (5,4), (5,3), (4,3), (3,3),(3,4), (3,5), (4,5), 
-                    (4,4)]
-
-safe_places=[ (1,4), (2,2), (2,6), (4,1), (4,4), (4,7), (6,2), (6,6), (7,4) ]
-
-home_places=[(4,8), (8,4), (4,0), (0,4)]
-
-COLOURS={
-    "BLACK" : (0, 0, 0) ,
-    "GRAY" : (180, 180, 180) ,
-    "WHITE" : (255, 255, 255) ,
-    "LIGHT_BLUE" : (102, 204, 255) ,
-    "RED" : (255, 0, 0) ,
-    "BLUE" : (30, 144, 255) ,
-    "GREEN" : (0, 255, 0) ,
-    "YELLOW" : (255, 255, 0) ,
-    "LIGHT_RED" : (255, 102, 102) ,
-    "LIGHT_GREEN" : (26, 255, 140) ,
-    "LIGHT_YELLOW" : (255, 255, 102)
-}
-
-
-def selectPath(color):
-    """ returns the path based on the given color(index) input """
-
-    if(color==0):
-        return redPath
-    elif(color==1):
-        return bluePath
-    elif(color==2):
-        return yellowPath
-    elif(color==3):
-        return greenPath
