+++ path.py
@@ -1,0 +1,15 @@
+paths = [[(0,4), (1,4), (1,3), (1,2), (1,1), (2,1), (3,1), (4,1), (5,1), (6,1), (7,1), (7,2), (7,3), (7,4), (7,5), (7,6), (7,7), (6,7), (5,7), (4,7), (3,7), (2,7), (1,7), (1,6), (1,5), 
+          (2,6), (3,6), (4,6), (5,6), (6,6), (6,5), (6,4), (6,3), (6,2), (5,2), (4,2), (3,2), (2,2), (2,3), (2,4), (2,5),  (3,5), (4,5), (5,5), (5,4), (5,3), (4,3), (3,3),(3,4), (4,4)],
+		
+		 [(4,0), (4,1), (5,1), (6,1), (7,1), (7,2), (7,3), (7,4), (7,5), (7,6), (7,7), (6,7), (5,7), (4,7), (3,7), (2,7), (1,7), (1,6), (1,5), (1,4), (1,3), (1,2), (1,1), (2,1), (3,1), 
+		  (2,2), (2,3), (2,4), (2,5), (2,6), (3,6), (4,6), (5,6), (6,6), (6,5), (6,4), (6,3), (6,2), (5,2), (4,2), (3,2), (3,3),(3,4), (3,5), (4,5), (5,5), (5,4), (5,3), (4,3), (4,4)],
+
+		 [(8,4), (7,4), (7,5), (7,6), (7,7), (6,7), (5,7), (4,7), (3,7), (2,7), (1,7), (1,6), (1,5), (1,4), (1,3), (1,2), (1,1), (2,1), (3,1), (4,1), (5,1), (6,1), (7,1), (7,2), (7,3), 
+          (6,2), (5,2), (4,2), (3,2), (2,2), (2,3), (2,4), (2,5), (2,6), (3,6), (4,6), (5,6), (6,6), (6,5), (6,4), (6,3), (5,3), (4,3),(3,3),(3,4), (3,5), (4,5), (5,5), (5,4), (4,4)],
+			
+		 [(4,8), (4,7), (3,7), (2,7), (1,7), (1,6), (1,5), (1,4), (1,3), (1,2), (1,1), (2,1), (3,1), (4,1), (5,1), (6,1), (7,1), (7,2), (7,3), (7,4), (7,5), (7,6), (7,7), (6,7), (5,7),
+          (6,6), (6,5), (6,4), (6,3), (6,2), (5,2), (4,2), (3,2), (2,2), (2,3), (2,4), (2,5), (2,6), (3,6), (4,6), (5,6), (5,5), (5,4), (5,3), (4,3), (3,3),(3,4), (3,5), (4,5), (4,4)] 
+		]
+
+def getPath(color):
+    return paths[color]
